Updates for OSAP
Removed old files that were used to test login features. I Also updated
the database with a new table to allow for course registration. I also
allowed for students to register for there courses. Improvements were
made to the login form to allow javascript popup in the event of invalid
login